requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.121.2
python-multipart==0.0.20
h11==0.16.0
httptools==0.7.1
idna==3.11
psycopg==3.2.12
psycopg-binary==3.2.12
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1

# LLM providers
openai==1.58.1
anthropic==0.39.0

# Testing
pytest==7.4.3
httpx==0.25.2

# Development tools
black==23.12.1
isort==5.13.2
flake8==6.1.0

# JWT for auth
pyjwt==2.9.0
cryptography==44.0.0

# Pronunciation scoring
phonemizer==3.2.1
numpy==1.26.4

# Payment processing
stripe==11.2.0
